requirements.txt
pytorch_lightning==1.7.0
torch==1.7.1 
librosa==0.8.0
museval==0.4.0
h5py==2.10.0
pytorch_lightning==1.2.1
numpy==1.18.5
torchlibrosa==0.0.9
matplotlib==3.3.4
musdb==0.4.0
museval==0.4.0
samplerate==0.1.0
audiomentations==0.34.1
hydra-core==1.3.2

python-dotenv==1.0.1
pyyaml==6.0.1
rich==13.7.1
h5py==3.10.0
museval==0.4.1